Refactor make meme to generator
